requirements.txt
pysimplegui
Pillow